Sept 2025 update (#13)
* Updated Skills

* Re ordered the components

* Updated Experience
