+++ package.json
@@ -20,7 +20,7 @@
     "clsx": "^2.1.1",
     "framer-motion": "^11.2.13",
     "lucide-react": "^0.400.0",
-    "next": "14.2.4",
+    "next": "^15.5.3",
     "next-themes": "^0.3.0",
     "react": "^18",
     "react-countup": "^6.5.3",
